[#702] Add utility: list-style
After #696 we’ll have no a-list-reset class. Let’s replace the functionality with a more standardised utility class list to specify the list-style-type. Default values could be none, disc, and decimal.

Remove padding & margin from all ol and ul elements at the base level — these can be specified for content like rich-text later, on a per-project basis.

Find all instances of a-list-reset in the docs, and replace them with the new utility class.

If there are any instances in our docs where we have a list without .a-list-reset, investigate. If we need the styling to remaining, perhaps create a rich-text component that applies the required styling for “content” lists e.g.

.c-rich-text {
  ul {
    padding-left: size.get('l');
  }
}

https://github.com/bitcrowd/bitstyles/issues/702